Refactor API routes and tags for improved organization
This commit refactors the API routes by removing outdated common and entry routes, and reorganizing the remaining routes into more specific categories such as matches, events, fixtures, leagues, and teams. Additionally, the tags associated with the API have been updated to reflect this new structure, enhancing clarity and maintainability. This change adheres to best practices for API design and improves the overall organization of the codebase.
