+++ src/index.ts
@@ -38,14 +38,22 @@ initConnections().then(() => {
                         description: 'API for LetLetMe application',
                     },
                     tags: [
-                        { name: 'common', description: 'Common endpoints' },
-                        { name: 'entry', description: 'Match entry endpoints' },
+                        { name: 'system', description: 'System endpoints' },
+                        { name: 'matches', description: 'Match endpoints' },
+                        { name: 'events', description: 'Event endpoints' },
+                        { name: 'fixtures', description: 'Fixture endpoints' },
+                        { name: 'leagues', description: 'League endpoints' },
                         { name: 'live', description: 'Live match endpoints' },
-                        { name: 'player', description: 'Player endpoints' },
-                        { name: 'stat', description: 'Statistics endpoints' },
-                        { name: 'summary', description: 'Summary endpoints' },
+                        { name: 'notices', description: 'Notice endpoints' },
+                        { name: 'players', description: 'Player endpoints' },
                         {
-                            name: 'tournament',
+                            name: 'statistics',
+                            description: 'Statistics endpoints',
+                        },
+                        { name: 'summaries', description: 'Summary endpoints' },
+                        { name: 'teams', description: 'Team endpoints' },
+                        {
+                            name: 'tournaments',
                             description: 'Tournament endpoints',
                         },
                     ],
