--- src/routes/v1/summary/index.ts
@@ -1,499 +1,0 @@
-import { Elysia, t } from 'elysia';
-import { HttpStatusCode } from 'elysia-http-status-code';
-
-export const summaryRoutes = new Elysia({ prefix: '/summary' })
-    .use(HttpStatusCode())
-    // Get tournament summary
-    .get(
-        '/tournament/:id',
-        async ({ params }) => {
-            // Mock response - to be implemented with actual database queries
-            return {
-                status: 'success',
-                data: {
-                    tournamentId: params.id,
-                    name: 'Summer Championship 2023',
-                    startDate: '2023-06-15',
-                    endDate: '2023-06-30',
-                    status: 'in_progress',
-                    currentRound: 3,
-                    totalRounds: 5,
-                    matchesCompleted: 48,
-                    matchesRemaining: 16,
-                    participantsCount: 64,
-                    topPerformers: [
-                        {
-                            playerId: 'p1',
-                            playerName: 'John Doe',
-                            status: 'active',
-                            statistics: {
-                                wins: 3,
-                                setsWon: 6,
-                                averageScore: 21.5,
-                            },
-                        },
-                        {
-                            playerId: 'p2',
-                            playerName: 'Jane Smith',
-                            status: 'active',
-                            statistics: {
-                                wins: 3,
-                                setsWon: 6,
-                                averageScore: 20.8,
-                            },
-                        },
-                        {
-                            playerId: 'p3',
-                            playerName: 'Mike Johnson',
-                            status: 'active',
-                            statistics: {
-                                wins: 3,
-                                setsWon: 6,
-                                averageScore: 19.2,
-                            },
-                        },
-                    ],
-                    upcomingMatches: [
-                        {
-                            matchId: 'm1',
-                            round: 3,
-                            player1: {
-                                id: 'p1',
-                                name: 'John Doe',
-                            },
-                            player2: {
-                                id: 'p8',
-                                name: 'Robert Brown',
-                            },
-                            scheduledTime: new Date(
-                                Date.now() + 3600000,
-                            ).toISOString(),
-                            court: 1,
-                        },
-                        {
-                            matchId: 'm2',
-                            round: 3,
-                            player1: {
-                                id: 'p2',
-                                name: 'Jane Smith',
-                            },
-                            player2: {
-                                id: 'p7',
-                                name: 'Emma Wilson',
-                            },
-                            scheduledTime: new Date(
-                                Date.now() + 7200000,
-                            ).toISOString(),
-                            court: 2,
-                        },
-                    ],
-                    recentResults: [
-                        {
-                            matchId: 'm10',
-                            round: 2,
-                            player1: {
-                                id: 'p1',
-                                name: 'John Doe',
-                            },
-                            player2: {
-                                id: 'p16',
-                                name: 'Thomas Clark',
-                            },
-                            score: [21, 18],
-                            winner: 'p1',
-                            completedAt: new Date(
-                                Date.now() - 3600000,
-                            ).toISOString(),
-                        },
-                        {
-                            matchId: 'm11',
-                            round: 2,
-                            player1: {
-                                id: 'p2',
-                                name: 'Jane Smith',
-                            },
-                            player2: {
-                                id: 'p15',
-                                name: 'Linda Lee',
-                            },
-                            score: [21, 15, 21, 18],
-                            winner: 'p2',
-                            completedAt: new Date(
-                                Date.now() - 7200000,
-                            ).toISOString(),
-                        },
-                    ],
-                },
-            };
-        },
-        {
-            detail: {
-                tags: ['summary'],
-                summary: 'Get tournament summary',
-                description: 'Returns a comprehensive summary of a tournament',
-                responses: {
-                    200: {
-                        description: 'Tournament summary',
-                    },
-                    404: {
-                        description: 'Tournament not found',
-                    },
-                },
-            },
-            params: t.Object({
-                id: t.String(),
-            }),
-        },
-    )
-    // Get dashboard summary
-    .get(
-        '/dashboard',
-        async ({ query }) => {
-            // Mock response - to be implemented with actual database queries
-            const date =
-                (query.date as string) ||
-                new Date().toISOString().split('T')[0];
-
-            return {
-                status: 'success',
-                data: {
-                    date,
-                    activeTournaments: 3,
-                    upcomingTournaments: 5,
-                    recentlyCompletedTournaments: 2,
-                    totalRegisteredPlayers: 256,
-                    activeMatches: 12,
-                    matchesCompletedToday: 24,
-                    upcomingMatchesToday: 36,
-                    tournaments: [
-                        {
-                            id: 't1',
-                            name: 'Summer Championship 2023',
-                            status: 'active',
-                            currentRound: 3,
-                            progress: 0.65, // 65% complete
-                            matchesCompleted: 42,
-                            matchesRemaining: 22,
-                        },
-                        {
-                            id: 't2',
-                            name: 'Junior Open 2023',
-                            status: 'active',
-                            currentRound: 2,
-                            progress: 0.4, // 40% complete
-                            matchesCompleted: 32,
-                            matchesRemaining: 48,
-                        },
-                        {
-                            id: 't3',
-                            name: 'Masters Series 2023',
-                            status: 'active',
-                            currentRound: 4,
-                            progress: 0.8, // 80% complete
-                            matchesCompleted: 60,
-                            matchesRemaining: 15,
-                        },
-                        {
-                            id: 't4',
-                            name: 'Fall Classic 2023',
-                            status: 'upcoming',
-                            startDate: '2023-09-15',
-                            registrationOpen: true,
-                            registeredPlayers: 48,
-                        },
-                    ],
-                    topRankedPlayers: [
-                        {
-                            id: 'p1',
-                            name: 'John Doe',
-                            country: 'US',
-                            ranking: 1,
-                            points: 5000,
-                        },
-                        {
-                            id: 'p2',
-                            name: 'Jane Smith',
-                            country: 'UK',
-                            ranking: 2,
-                            points: 4850,
-                        },
-                        {
-                            id: 'p3',
-                            name: 'Mike Johnson',
-                            country: 'CA',
-                            ranking: 3,
-                            points: 4720,
-                        },
-                    ],
-                    recentResults: [
-                        {
-                            matchId: 'm1',
-                            tournamentId: 't1',
-                            tournamentName: 'Summer Championship 2023',
-                            player1: 'John Doe',
-                            player2: 'Robert Brown',
-                            score: [21, 19],
-                            completedAt: new Date(
-                                Date.now() - 3600000,
-                            ).toISOString(),
-                        },
-                        {
-                            matchId: 'm2',
-                            tournamentId: 't2',
-                            tournamentName: 'Junior Open 2023',
-                            player1: 'Emma Wilson',
-                            player2: 'Thomas Clark',
-                            score: [21, 18, 18, 21, 21, 15],
-                            completedAt: new Date(
-                                Date.now() - 7200000,
-                            ).toISOString(),
-                        },
-                    ],
-                },
-            };
-        },
-        {
-            detail: {
-                tags: ['summary'],
-                summary: 'Get dashboard summary',
-                description:
-                    'Returns a comprehensive summary for the dashboard',
-                responses: {
-                    200: {
-                        description: 'Dashboard summary',
-                    },
-                },
-            },
-            query: t.Object({
-                date: t.Optional(t.String()),
-            }),
-        },
-    )
-    // Get match summary
-    .get(
-        '/match/:id',
-        async ({ params }) => {
-            // Mock response - to be implemented with actual database queries
-            return {
-                status: 'success',
-                data: {
-                    matchId: params.id,
-                    tournamentId: 't1',
-                    tournamentName: 'Summer Championship 2023',
-                    round: 3,
-                    court: 2,
-                    player1: {
-                        id: 'p1',
-                        name: 'John Doe',
-                        country: 'US',
-                        ranking: 1,
-                        seedNumber: 1,
-                        tournamentStats: {
-                            matchesPlayed: 3,
-                            matchesWon: 3,
-                            setsWon: 6,
-                            setsLost: 1,
-                        },
-                    },
-                    player2: {
-                        id: 'p8',
-                        name: 'Robert Brown',
-                        country: 'CA',
-                        ranking: 12,
-                        seedNumber: 8,
-                        tournamentStats: {
-                            matchesPlayed: 3,
-                            matchesWon: 3,
-                            setsWon: 6,
-                            setsLost: 2,
-                        },
-                    },
-                    status: 'completed',
-                    score: [21, 18, 19, 21, 21, 15],
-                    winner: 'p1',
-                    duration: 68, // minutes
-                    startTime: new Date(Date.now() - 5400000).toISOString(), // 1.5 hours ago
-                    endTime: new Date(Date.now() - 1500000).toISOString(), // 25 minutes ago
-                    statistics: {
-                        player1: {
-                            aces: 5,
-                            doubleFaults: 2,
-                            winnersCount: 28,
-                            errorsCount: 15,
-                            firstServePercentage: 0.75,
-                            breakPointsConverted: 3,
-                            breakPointsTotal: 5,
-                        },
-                        player2: {
-                            aces: 3,
-                            doubleFaults: 4,
-                            winnersCount: 22,
-                            errorsCount: 18,
-                            firstServePercentage: 0.68,
-                            breakPointsConverted: 2,
-                            breakPointsTotal: 6,
-                        },
-                    },
-                    pointByPoint: [
-                        {
-                            set: 1,
-                            points: [
-                                { server: 'p1', winner: 'p1', score: [1, 0] },
-                                { server: 'p1', winner: 'p1', score: [2, 0] },
-                                // More points would be listed here
-                            ],
-                        },
-                        // More sets would be listed here
-                    ],
-                    nextMatch: {
-                        id: 'mNext',
-                        round: 4,
-                        player1: {
-                            id: 'p1',
-                            name: 'John Doe',
-                        },
-                        player2: {
-                            id: 'p4',
-                            name: 'Sarah Williams',
-                        },
-                        scheduledTime: new Date(
-                            Date.now() + 86400000,
-                        ).toISOString(), // tomorrow
-                    },
-                },
-            };
-        },
-        {
-            detail: {
-                tags: ['summary'],
-                summary: 'Get match summary',
-                description: 'Returns a detailed summary of a match',
-                responses: {
-                    200: {
-                        description: 'Match summary',
-                    },
-                    404: {
-                        description: 'Match not found',
-                    },
-                },
-            },
-            params: t.Object({
-                id: t.String(),
-            }),
-        },
-    )
-    // Get player summary
-    .get(
-        '/player/:id',
-        async ({ params, query }) => {
-            // Mock response - to be implemented with actual database queries
-            const tournamentId = query.tournamentId as string | undefined;
-
-            return {
-                status: 'success',
-                data: {
-                    playerId: params.id,
-                    name: 'John Doe',
-                    country: 'US',
-                    ranking: 1,
-                    points: 5000,
-                    tournaments: {
-                        total: 45,
-                        won: 12,
-                        runnerUp: 8,
-                        current: tournamentId
-                            ? {
-                                  id: tournamentId,
-                                  name: 'Summer Championship 2023',
-                                  status: 'active',
-                                  round: 3,
-                                  stats: {
-                                      matchesPlayed: 3,
-                                      matchesWon: 3,
-                                      setsWon: 6,
-                                      setsLost: 1,
-                                  },
-                                  nextMatch: {
-                                      id: 'm1',
-                                      round: 3,
-                                      opponent: {
-                                          id: 'p8',
-                                          name: 'Robert Brown',
-                                          ranking: 12,
-                                      },
-                                      scheduledTime: new Date(
-                                          Date.now() + 3600000,
-                                      ).toISOString(),
-                                      court: 1,
-                                  },
-                              }
-                            : null,
-                    },
-                    recentForm: [
-                        {
-                            result: 'W',
-                            tournamentId: 't1',
-                            tournamentName: 'Summer Championship 2023',
-                            opponentId: 'p16',
-                            opponentName: 'Thomas Clark',
-                            score: [21, 18],
-                            date: new Date(Date.now() - 86400000).toISOString(), // yesterday
-                        },
-                        {
-                            result: 'W',
-                            tournamentId: 't1',
-                            tournamentName: 'Summer Championship 2023',
-                            opponentId: 'p32',
-                            opponentName: 'Michael Lee',
-                            score: [21, 15],
-                            date: new Date(
-                                Date.now() - 172800000,
-                            ).toISOString(), // 2 days ago
-                        },
-                        {
-                            result: 'W',
-                            tournamentId: 't0',
-                            tournamentName: 'Spring Open 2023',
-                            opponentId: 'p2',
-                            opponentName: 'Jane Smith',
-                            score: [19, 21, 21, 18, 21, 15],
-                            date: new Date(
-                                Date.now() - 2592000000,
-                            ).toISOString(), // 30 days ago
-                        },
-                    ],
-                    seasonStats: {
-                        matchesPlayed: 65,
-                        matchesWon: 52,
-                        matchesLost: 13,
-                        winRate: 0.8,
-                        tournamentsPlayed: 18,
-                        tournamentsWon: 6,
-                        averageScore: 19.8,
-                        highestScore: 29,
-                    },
-                },
-            };
-        },
-        {
-            detail: {
-                tags: ['summary'],
-                summary: 'Get player summary',
-                description: 'Returns a comprehensive summary for a player',
-                responses: {
-                    200: {
-                        description: 'Player summary',
-                    },
-                    404: {
-                        description: 'Player not found',
-                    },
-                },
-            },
-            params: t.Object({
-                id: t.String(),
-            }),
-            query: t.Object({
-                tournamentId: t.Optional(t.String()),
-            }),
-        },
-    );
